Safari-611.1.10/LayoutTests/editing/secure-input/removed-password-input-expected.txt

Verify that removing a password input from DOM tree disables secure input state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


A password input is focused:
PASS testRunner.secureEventInputIsEnabled is true

After deleting the input:
FAIL testRunner.secureEventInputIsEnabled should be false. Was true.
PASS successfullyParsed is true

TEST COMPLETE
